+++ src/components/DataDisplay/ListAvatar/index.ts
@@ -1,0 +1,4 @@
+import ListAvatar from "./ListAvatar";
+
+export default ListAvatar;
+export * from "./ListAvatar";