+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # [Versions](https://github.com/Tracktor/design-system/releases)
 
-## v2.1.4
-- **[fix]** - time line tag margin with very long title
+## v2.2.0
+- **[feat]** - add new component `ListAvatar`
+- **[fix]** - `Timeline` active style